feat:service인터페이스화 및 불필요하게 전체 가져오는 import수정
