+++ src/main/java/com/flab/tiple/ticket/repository/TicketReservationRepository.java
@@ -1,12 +1,10 @@
 package com.flab.tiple.ticket.repository;
 
 import java.util.List;
-import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.flab.tiple.ticket.domain.TicketReservation;
-import com.flab.tiple.ticket.enums.TicketReservationStatus;
 
 public interface TicketReservationRepository extends JpaRepository<TicketReservation, Long> {
